requirements.txt
Flask==2.3.3
pymongo==4.5.0
transformers==4.33.2
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
nltk==3.8.1
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.3.7
Jinja2==3.1.2
dnspython==2.4.2
tokenizers==0.13.3
huggingface-hub==0.16.4
safetensors==0.3.3
regex==2023.8.8
click==8.1.7
itsdangerous==2.1.2
MarkupSafe==2.1.3
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
urllib3==2.0.4
joblib==1.3.2
scipy==1.11.2
threadpoolctl==3.2.0
packaging==23.1
filelock==3.12.4
fsspec==2023.9.2
pyyaml==6.0.1
typing-extensions==4.7.1 